+++ bundlers/TypeScriptHelper.js
@@ -7,8 +7,8 @@ var BrowserMultiPack = require("./browser/BrowserMultiPack");
  */
 var TypeScriptHelper;
 (function (TypeScriptHelper) {
-    // copied from TypeScript package: typescript/lib/tsserver.js
-    // used by compiled TypeScript code that utilizes 'extends' in classes, async functions, and decorators
+    /** copied from TypeScript package: typescript/lib/tsserver.js
+     * used by compiled TypeScript code that utilizes 'extends' in classes, async functions, and decorators */
     TypeScriptHelper.staticHelpers = {
         extendsHelper: "\nvar __extends = (this && this.__extends) || function (d, b) {\n  for (var p in b) if (b.hasOwnProperty(p)) d[p] = b[p];\n  function __() { this.constructor = d; }\n  d.prototype = b === null ? Object.create(b) : (__.prototype = b.prototype, new __());\n};",
         decorateHelper: "\nvar __decorate = (this && this.__decorate) || function (decorators, target, key, desc) {\n  var c = arguments.length, r = c < 3 ? target : desc === null ? desc = Object.getOwnPropertyDescriptor(target, key) : desc, d;\n  if (typeof Reflect === \"object\" && typeof Reflect.decorate === \"function\") r = Reflect.decorate(decorators, target, key, desc);\n  else for (var i = decorators.length - 1; i >= 0; i--) if (d = decorators[i]) r = (c < 3 ? d(r) : c > 3 ? d(target, key, r) : d(target, key)) || r;\n  return c > 3 && r && Object.defineProperty(target, key, r), r;\n};",
@@ -18,8 +18,7 @@ var TypeScriptHelper;
     };
     /** NOTE: typescript (i.e. 'tsc') must be installed and available via the command line.
      * Example: "tsc -t ES5 -m commonjs --preserveConstEnums --forceConsistentCasingInFileNames --noEmitHelpers " + projectRelativeSrcPath
-     * @param projRelativeSrcPath
-     * @param callback
+     * @param tscCmd the typescript compiler command to execute
      */
     function compileTypeScriptFile(tscCmd) {
         var dfd = Q.defer();
@@ -42,7 +41,7 @@ var TypeScriptHelper;
     }
     TypeScriptHelper.compileTypeScriptFile = compileTypeScriptFile;
     /** Get the 'prelude.js' source string used by browser-pack, insert the TypeScript static helpers (required for 'extends', annotations, and other TypeScript features) into it, and return it
-     * @param [includeUsageComment=false] whether to include a comment in the source string explaining why the TypeScript static helpers are inserted
+     * @param includeUsageComment optional (default: true), whether to include a comment in the source string explaining why the TypeScript static helpers are inserted
      */
     function createPreludeStringWithTypeScriptHelpers(includeUsageComment) {
         var preludeSrc = BrowserMultiPack.getPreludeSrc();