+++ test/test-proj/App.js
@@ -1,0 +1,7 @@
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
+var HelperUtil = require("./HelperUtil");
+function main() {
+    console.log(HelperUtil.App.name + "@" + HelperUtil.App.version);
+}
+main();