+++ test/test-proj/App.ts
@@ -1,0 +1,7 @@
+﻿import HelperUtil = require("./HelperUtil");
+
+function main() {
+    console.log(HelperUtil.App.name + "@" + HelperUtil.App.version);
+}
+
+main();