+++ test/test-proj/HelperUtil.js
@@ -1,0 +1,5 @@
+"use strict";
+var HelperUtil;
+(function (HelperUtil) {
+})(HelperUtil || (HelperUtil = {}));
+module.exports = HelperUtil;