+++ test/test-proj/HelperUtil.ts
@@ -1,0 +1,19 @@
+﻿import fs = require("fs");
+
+module HelperUtil {
+
+    interface PackageLike {
+        name: string;
+        version: string;
+        description?: string;
+        dependencies?: { [prop: string]: string };
+        devDependencies?: { [prop: string]: string };
+        [prop: string]: any
+    }
+
+
+    export var App: PackageLike;
+
+}
+
+export = HelperUtil;