+++ utils/StringUtil.ts
@@ -2,6 +2,15 @@
 module StringUtil {
 
 
+    /** Count '\n' characters in the given string
+     */
+    export function countNewlines(src: string) {
+        if (!src) return 0;
+        var newlines = src.match(/\n/g);
+        return (newlines != null ? newlines.length : 0);
+    }
+
+
     /** Add optional prefix and suffix strings to an array of strings as well as optional first line prefix and last line prefix to the first and last line
      * @return the input 'lines', now modified
      */